requirements.txt
pandas 
numpy 
matplotlib 
seaborn 
scikit-learn 
scipy
mlflow
python-box
PyYAML
tqdm
ensure
joblib
types-PyYAML
Flask
Flask-Cors
python-dotenv
# -e .